Moved TsRepository to interfaces.
